Move ScyllaDB interaction from tx-details-storage to database crate, add necessary implementation to keep the code consistent
