+++ database/Cargo.toml
@@ -23,7 +23,11 @@ sqlx = { version = "0.8.2", features = [
     "runtime-tokio-native-tls",
     "postgres",
     "migrate",
-    "bigdecimal"] }
+    "bigdecimal",
+] }
+num-bigint = "0.3.3"
+num-traits = "0.2.19"
+scylla = { version = "0.15.1", features = ["ssl", "full-serialization"] }
 
 configuration.workspace = true
 readnode-primitives.workspace = true
