+++ database/src/scylla/mod.rs
@@ -1,0 +1,1 @@
+pub mod tx_indexer;
