+++ tx-indexer/Cargo.toml
@@ -31,6 +31,7 @@ tokio-retry = "0.3"
 tokio-stream = "0.1.12"
 tracing = "0.1.34"
 
+database.workspace = true
 cache-storage.workspace = true
 configuration.workspace = true
 readnode-primitives.workspace = true
